split settings package
